Flatten lists into Hunter
