+++ test/hunter/list_test.exs
@@ -12,7 +12,7 @@ defmodule Hunter.ListTest do
       respond_with_fixture(conn, "list", wrap: :list)
     end)
 
-    assert [%List{id: "12249", title: "Friends"}] = List.lists(@conn)
+    assert [%List{id: "12249", title: "Friends"}] = Hunter.lists(@conn)
   end
 
   test "returns a single list" do
@@ -22,7 +22,7 @@ defmodule Hunter.ListTest do
       respond_with_fixture(conn, "list")
     end)
 
-    assert %List{id: "12249", replies_policy: "followed"} = List.list(@conn, 12_249)
+    assert %List{id: "12249", replies_policy: "followed"} = Hunter.list(@conn, 12_249)
   end
 
   test "creates a list with a JSON body" do
@@ -34,7 +34,7 @@ defmodule Hunter.ListTest do
     end)
 
     assert %List{title: "Friends", replies_policy: "followed"} =
-             List.create_list(@conn, "Friends", replies_policy: "followed")
+             Hunter.create_list(@conn, "Friends", replies_policy: "followed")
   end
 
   test "updates a list" do
@@ -46,7 +46,7 @@ defmodule Hunter.ListTest do
     end)
 
     assert %List{id: "12249"} =
-             List.update_list(@conn, 12_249, title: "Close friends", exclusive: true)
+             Hunter.update_list(@conn, 12_249, title: "Close friends", exclusive: true)
   end
 
   test "destroys a list" do
@@ -56,7 +56,7 @@ defmodule Hunter.ListTest do
       respond_with(conn, %{})
     end)
 
-    assert List.destroy_list(@conn, 12_249) == true
+    assert Hunter.destroy_list(@conn, 12_249) == true
   end
 
   test "returns accounts in a list" do
@@ -68,7 +68,7 @@ defmodule Hunter.ListTest do
     end)
 
     assert [%Hunter.Account{username: "milmazz"}] =
-             List.list_accounts(@conn, 12_249, limit: 1)
+             Hunter.list_accounts(@conn, 12_249, limit: 1)
   end
 
   test "adds accounts to a list with a JSON body" do
@@ -79,7 +79,7 @@ defmodule Hunter.ListTest do
       respond_with(conn, %{})
     end)
 
-    assert List.add_accounts_to_list(@conn, 12_249, [8039]) == true
+    assert Hunter.add_accounts_to_list(@conn, 12_249, [8039]) == true
   end
 
   test "removes accounts from a list via query params" do
@@ -90,7 +90,7 @@ defmodule Hunter.ListTest do
       respond_with(conn, %{})
     end)
 
-    assert List.remove_accounts_from_list(@conn, 12_249, [8039]) == true
+    assert Hunter.remove_accounts_from_list(@conn, 12_249, [8039]) == true
   end
 
   test "returns lists containing a given account" do
@@ -100,7 +100,7 @@ defmodule Hunter.ListTest do
       respond_with_fixture(conn, "list", wrap: :list)
     end)
 
-    assert [%List{title: "Friends"}] = List.account_lists(@conn, 8039)
+    assert [%List{title: "Friends"}] = Hunter.account_lists(@conn, 8039)
   end
 
   test "API errors raise Hunter.Error" do
@@ -108,6 +108,6 @@ defmodule Hunter.ListTest do
       respond_with(conn, %{error: "Record not found"}, 404)
     end)
 
-    assert_raise Hunter.Error, fn -> List.list(@conn, 0) end
+    assert_raise Hunter.Error, fn -> Hunter.list(@conn, 0) end
   end
 end
